Reorganize CodeFixers
